// src/main/resources/angular/src/app/api/model/guildCard.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Guild } from './guild';
import { GuildMember } from './guildMember';


export interface GuildCard { 
    cardId: number;
    guild: Guild;
    cardAssignationGuildDate?: string;
    cardAssignationMemberDate?: string;
    cardCreatedDate?: string;
    initialBalance?: number;
    guildMember?: GuildMember;
    balance?: number;
    cardNumber?: number;
	guildCardStatusId?: number;
	previousDisabledStatus?: number;
	deletedDate?: string;
	expiredParkingDate?: string;
}