// src/main/resources/angular/src/app/api/model/guildCardCreate.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface GuildCardCreate {
	cardTo: number;
	cardFrom: number;
	guild?: number;
	initialBalance?: number;
}